+++ soloconnect/soloconnect/settings.py
@@ -99,7 +99,7 @@ AUTH_PASSWORD_VALIDATORS = [
     },
 ]
 
-AUTH_USER_MODEL = 'user_profile.CustomUser'
+AUTH_USER_MODEL = "user_profile.CustomUser"
 
 # Internationalization
 # https://docs.djangoproject.com/en/4.2/topics/i18n/
